client:라우터 추가 및 MVP 형태로 리팩토링
