+++ client/src/App.tsx
@@ -1,22 +1,17 @@
-import React, { FC, useEffect } from 'react';
-import { CardRequest, CardResponse } from './proto/poker_pb';
-import { PokerServiceClient } from './proto/PokerServiceClientPb';
+import React, { FC } from 'react';
+import { BrowserRouter, Route, Routes } from 'react-router-dom';
+import Home from './presenters/Home';
+import PokerRoom from './presenters/PokerRoom';
 
-const client = new PokerServiceClient('http://localhost:8080');
 const App: FC = () => {
-  useEffect(() => {
-    const req = new CardRequest();
-    req.setUsername('ek');
-    req.setUid('132');
-    req.setScore(3);
-
-    const cardStream = client.cardStream(req, {});
-    cardStream.on('data', (chunk: CardResponse) => {
-      const msg = chunk.toObject();
-      console.log(msg);
-    });
-  }, []);
-  return <div className="App">hello world</div>;
+  return (
+    <BrowserRouter>
+      <Routes>
+        <Route path="/" element={<Home />} />
+        <Route path="poker-room" element={<PokerRoom />} />
+      </Routes>
+    </BrowserRouter>
+  );
 };
 
 export default App;
